fix: simplify mutation observer implementation to reduce slow down.
It seems to be quicker to run ahocorasick on the document body than to isolate altered nodes and run ahocorasick on all of them. Noticeably reduces slow down on complex pages like youtube and when many new nodes are added on user input e.g. suggestions dropdown when typing in Google search bar.